Fix memory leak where ZBarImage was not freed after use
I left the application doing the barcode scanning for a while and
noticed that the amount of memory allocated by zbar_image_create() was
constantly increasing. Adding `Drop` for `ZBarImage` seemed to be
enough, but caused double-free errors in zbar_sumbol.
Turns out that `Vec::from_raw_parts` takes ownership of the data
provided to it, and frees it itself after it goes out of scope.
zbar_symbol was not aware of that, and tried to free the memory by
itself when the ZBarImage was destroyed.
Using `slice::from_raw_parts` avoids this issue, as it does not take
the ownership of data. After that `to_vec()` copies the data, which
we can manage as we please.